LayoutTests/platform/qt-mac/fast/images/zoomed-img-size-expected.txt

Ideally, all ovals below should be reported as having width=37, height=33. Currently rounding prevents us from doing this.


Zoom 1% PASS: 100x100 close enough to 37x33


Zoom 2% PASS: 50x50 close enough to 37x33


Zoom 3% PASS: 33x33 close enough to 37x33


Zoom 4% PASS: 25x25 close enough to 37x33


Zoom 5% PASS: 20x20 close enough to 37x33


Zoom 30% PASS: 36x30 close enough to 37x33


Zoom 33% PASS: 36x30 close enough to 37x33


Zoom 50% PASS: 36x32 close enough to 37x33


Zoom 70% PASS: 35x32 close enough to 37x33


Zoom 100% PASS: 37x33


Zoom 111% PASS: 37x33


Zoom 150% PASS: 37x33


Zoom 333% PASS: 37x33


Zoom 400% PASS: 37x33


Zoom 1234% PASS: 37x33
